requirements.txt
# Core dependencies
flask==3.0.0
werkzeug==3.0.1
flask-cors==4.0.0
flask-limiter==3.5.0
flask-session==0.5.0
bcrypt==4.1.2
tencentcloud-sdk-python==3.0.1035
requests==2.31.0

# Cache and database
redis==5.0.1
psutil==5.9.6
pymysql==1.1.0
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
cryptography==41.0.7

# Configuration and serialization
pyyaml==6.0.1

# File type detection (optional, for enhanced security)
python-magic==0.4.27

# Performance monitoring and analysis
memory-profiler==0.61.0

# Development tools (optional)
# python-dotenv==1.0.0  # 已移除环境变量依赖，保留注释供参考
# watchdog==0.0       # File system monitoring